python_agents/requirements.txt
python-dateutil>=2.8.0
redis
psycopg2-binary
python-dotenv
google-generativeai
requests
asyncio
